refactor: 폴더 구조 변경
- 컴포넌트를 같은 관심사에 따라 폴더 구조를 가지도록 변경함
