+++ README.md
@@ -5,11 +5,13 @@
 ## 기능 설명
 
 ### 일정 추가
+
 상단의 + 버튼을 통해 일정 추가가 가능합니다. **제목, 날짜, 시작 시간, 종료 시간, 설명** 등의 데이터를 필요로 하고, 날짜와 제목 외에는 생략이 가능합니다.
 
 <img width="1440" alt="스크린샷 2025-01-12 오후 9 27 12" src="https://github.com/user-attachments/assets/769f116e-d798-4c98-963e-32fd30b86e51" />
 
 ### 일정 삭제, 수정
+
 <img width="1440" alt="스크린샷 2025-01-12 오후 9 29 06" src="https://github.com/user-attachments/assets/593f5114-00ad-4037-bd81-1311f5367b4c" />
 
 추가된 일정을 누르게 되면 상세 정보 모달이 뜨는데, 우측 버튼을 통해 삭제와 수정이 가능합니다.
