+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Calendar from '@/components/Calendar';
+import Calendar from '@/components/Calendar/Calendar';
 
 export default function Home() {
   return (
